requirements.txt
Django==4.1
gunicorn==21.2.0
redis==5.0.1
watchdog==3.0.0
django-redis==5.4.0
whitenoise==6.6.0
python-decouple==3.8
django-rest-framework==0.1.0
Faker==19.12.0